Moved ColorMapSeries to series folder.
