+++ ts/Series/Heatmap/HeatmapSeries.ts
@@ -25,7 +25,7 @@ import type { PointStateHoverOptions } from '../../Core/Series/PointOptions';
 import type { StatesOptionsKey } from '../../Core/Series/StatesOptions';
 import type SVGAttributes from '../../Core/Renderer/SVG/SVGAttributes';
 import Color from '../../Core/Color/Color.js';
-import ColorMapMixin from '../../Mixins/ColorMapSeries.js';
+import ColorMapMixin from '../ColorMapSeries.js';
 const { colorMapSeriesMixin } = ColorMapMixin;
 import HeatmapPoint from './HeatmapPoint.js';
 import LegendSymbolMixin from '../../Mixins/LegendSymbol.js';
