+++ ts/Series/Map/MapPoint.ts
@@ -22,7 +22,7 @@ import type PointerEvent from '../../Core/PointerEvent';
 import type { PointShortOptions } from '../../Core/Series/PointOptions';
 import type ScatterPoint from '../Scatter/ScatterPoint';
 import type SVGPath from '../../Core/Renderer/SVG/SVGPath';
-import ColorMapMixin from '../../Mixins/ColorMapSeries.js';
+import ColorMapMixin from '../ColorMapSeries.js';
 const { colorMapPointMixin } = ColorMapMixin;
 import SeriesRegistry from '../../Core/Series/SeriesRegistry.js';
 const {
